+++ packages/wonder-blocks-dropdown/src/components/check.tsx
@@ -1,11 +1,9 @@
 import * as React from "react";
 import {StyleSheet} from "aphrodite";
 
-import Color from "@khanacademy/wonder-blocks-color";
 import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
+import Spacing from "@khanacademy/wonder-blocks-spacing";
 import checkIcon from "@phosphor-icons/core/bold/check-bold.svg";
-
-const {offBlack, offBlack32, white} = Color;
 
 /**
  * Props describing the state of the OptionItem, shared by the checkbox
@@ -16,30 +14,17 @@ type CheckProps = {
     disabled: boolean;
     /** Whether option item is selected. */
     selected: boolean;
-    /** Whether option item is pressed. */
-    pressed: boolean;
-    /** Whether option item is hovered. */
-    hovered: boolean;
-    /** Whether option item is focused. */
-    focused: boolean;
 };
 
 /**
  * The check component used by OptionItem.
  */
 const Check = function (props: CheckProps): React.ReactElement {
-    const {disabled, selected, pressed, hovered, focused} = props;
+    const {selected} = props;
     return (
         <PhosphorIcon
             icon={checkIcon}
             size="small"
-            color={
-                disabled
-                    ? offBlack32
-                    : pressed || hovered || focused
-                    ? white
-                    : offBlack
-            }
             style={[styles.bounds, !selected && styles.hide]}
         />
     );
@@ -49,9 +34,11 @@ export default Check;
 
 const styles = StyleSheet.create({
     bounds: {
+        alignSelf: "center",
+        height: Spacing.medium_16,
         // Semantically, this are the constants for a small-sized icon
-        minHeight: 16,
-        minWidth: 16,
+        minHeight: Spacing.medium_16,
+        minWidth: Spacing.medium_16,
     },
 
     hide: {