+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 1.2.5
+
+### Patch Changes
+
+-   Updated dependencies [23ab9f8c]
+-   Updated dependencies [6df21f71]
+    -   @khanacademy/wonder-blocks-icon@4.0.1
+    -   @khanacademy/wonder-blocks-core@6.3.1
+    -   @khanacademy/wonder-blocks-clickable@4.0.12
+    -   @khanacademy/wonder-blocks-typography@2.1.10
+
 ## 1.2.4
 
 ### Patch Changes