+++ packages/wonder-blocks-layout/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-layout
 
+## 2.0.25
+
+### Patch Changes
+
+-   Updated dependencies [6df21f71]
+    -   @khanacademy/wonder-blocks-core@6.3.1
+
 ## 2.0.24
 
 ### Patch Changes