+++ packages/wonder-blocks-pill/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-pill
 
+## 1.0.15
+
+### Patch Changes
+
+-   Updated dependencies [6df21f71]
+    -   @khanacademy/wonder-blocks-core@6.3.1
+    -   @khanacademy/wonder-blocks-link@6.0.5
+    -   @khanacademy/wonder-blocks-clickable@4.0.12
+    -   @khanacademy/wonder-blocks-typography@2.1.10
+
 ## 1.0.14
 
 ### Patch Changes